fix: only apply extractErrors to JSONAPI serializers
