+++ tests/main/tests/integration/serializers/json-serializer-test.js
@@ -1229,134 +1229,6 @@ module('integration/serializer/json - JSONSerializer', function (hooks) {
     assert.deepEqual(json, expected, 'returned JSON is correct');
   });
 
-  test('extractErrors respects custom key mappings', function (assert) {
-    class Post extends Model {
-      @attr('string') title;
-      @attr('string') description;
-      @attr('string') anotherString;
-      @attr('string') content;
-      @attr('string') inHash;
-      @attr('string') notInHash;
-      @belongsTo('person', { async: false, inverse: 'posts' }) author;
-      @hasMany('comment', { inverse: null, async: true }) comments;
-    }
-    class Comment extends Model {
-      @attr('string') body;
-      @belongsTo('post', { inverse: null, async: true }) post;
-    }
-    class Person extends Model {
-      @hasMany('post', { async: false, inverse: 'author' }) posts;
-    }
-
-    this.owner.register('model:post', Post);
-    this.owner.register('model:comment', Comment);
-    this.owner.register('model:person', Person);
-    this.owner.register(
-      'serializer:post',
-      JSONSerializer.extend({
-        attrs: {
-          title: 'le_title',
-          comments: { key: 'my_comments' },
-        },
-      })
-    );
-
-    var payload = {
-      errors: [
-        {
-          source: { pointer: 'data/attributes/le_title' },
-          detail: 'title errors',
-        },
-        {
-          source: { pointer: 'data/attributes/my_comments' },
-          detail: 'comments errors',
-        },
-      ],
-    };
-
-    const store = this.owner.lookup('service:store');
-    var errors = store.serializerFor('post').extractErrors(store, store.modelFor('post'), payload);
-
-    assert.deepEqual(errors, {
-      title: ['title errors'],
-      comments: ['comments errors'],
-    });
-  });
-
-  test('extractErrors expects error information located on the errors property of payload', function (assert) {
-    class Post extends Model {
-      @attr('string') title;
-      @attr('string') description;
-      @attr('string') anotherString;
-      @attr('string') content;
-      @attr('string') inHash;
-      @attr('string') notInHash;
-      @belongsTo('person', { async: false, inverse: 'posts' }) author;
-      @hasMany('comment', { inverse: null, async: true }) comments;
-    }
-    class Comment extends Model {
-      @attr('string') body;
-      @belongsTo('post', { inverse: null, async: true }) post;
-    }
-    class Person extends Model {
-      @hasMany('post', { async: false, inverse: 'author' }) posts;
-    }
-
-    this.owner.register('model:post', Post);
-    this.owner.register('model:comment', Comment);
-    this.owner.register('model:person', Person);
-    this.owner.register('serializer:post', JSONSerializer.extend());
-
-    var payload = {
-      attributeWhichWillBeRemovedInExtractErrors: ['true'],
-      errors: [
-        {
-          source: { pointer: 'data/attributes/title' },
-          detail: 'title errors',
-        },
-      ],
-    };
-
-    const store = this.owner.lookup('service:store');
-    var errors = store.serializerFor('post').extractErrors(store, store.modelFor('post'), payload);
-
-    assert.deepEqual(errors, { title: ['title errors'] });
-  });
-
-  test('extractErrors leaves payload untouched if it has no errors property', function (assert) {
-    class Post extends Model {
-      @attr('string') title;
-      @attr('string') description;
-      @attr('string') anotherString;
-      @attr('string') content;
-      @attr('string') inHash;
-      @attr('string') notInHash;
-      @belongsTo('person', { async: false, inverse: 'posts' }) author;
-      @hasMany('comment', { inverse: null, async: true }) comments;
-    }
-    class Comment extends Model {
-      @attr('string') body;
-      @belongsTo('post', { inverse: null, async: true }) post;
-    }
-    class Person extends Model {
-      @hasMany('post', { async: false, inverse: 'author' }) posts;
-    }
-
-    this.owner.register('model:post', Post);
-    this.owner.register('model:comment', Comment);
-    this.owner.register('model:person', Person);
-    this.owner.register('serializer:post', JSONSerializer.extend());
-
-    var payload = {
-      untouchedSinceNoErrorsSiblingPresent: ['true'],
-    };
-
-    const store = this.owner.lookup('service:store');
-    var errors = store.serializerFor('post').extractErrors(store, store.modelFor('post'), payload);
-
-    assert.deepEqual(errors, { untouchedSinceNoErrorsSiblingPresent: ['true'] });
-  });
-
   test('normalizeResponse should extract meta using extractMeta', function (assert) {
     class Post extends Model {
       @attr('string') title;
